question_001_100/question_91_100/98_is_valid_BST_test.go
package question_91_100

import (
	"testing"
)

func Test_isValidBST(t *testing.T) {
	tests := []struct {
		name string
		root *TreeNode
		want bool
	}{
		{"test", nil, true},
		{"test", &TreeNode{Val: 1}, true},
		{"test", &TreeNode{Val: 1, Left: &TreeNode{Val: -1}, Right: &TreeNode{Val: 2}}, true},
		{"test", &TreeNode{Val: 2, Left: &TreeNode{Val: 1}, Right: &TreeNode{Val: 3}}, true},
		{"test", &TreeNode{Val: 1, Left: &TreeNode{Val: 1}, Right: &TreeNode{Val: 2}}, true},
		{"test", &TreeNode{Val: 5, Left: &TreeNode{Val: 1}, Right: &TreeNode{Val: 4, Left: &TreeNode{Val: 3}, Right: &TreeNode{Val: 6}}}, false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := isValidBST(tt.root); got != tt.want {
				t.Errorf("isValidBST() = %v, want %v", got, tt.want)
			}
		})
	}
}
